# full-stack-asset-transfer-guide/applications/rest-api/src/coconiko/coconiko_swagger.yaml
openapi: 3.0.1
info:
  title: Coconiko API
  version: 1.1.0
  description: |
    FabricネットワークにおけるユーザーとCoconikoのコインとNFTなどを管理するためのAPI。

    アプリからの呼び出しに最適化しており、ヘッダーに`X-API-Key`とログインしている`userId`を設定し、
    APIを呼び出す仕様となっています。
    - `X-Api-Key`で呼び出し元のシステムを識別しています
    - `userId`でログインしているユーザーを識別しています

    各APIの必要な認証情報は各APIの右側のロックアイコンで確認できます
    
    注意：「ユーザーID」と「アカウントID」が違います。
    - 「アカウントID」はFabricネットワーク上でユーザーを識別するIDです、新規ユーザーを登録する時に割り当てられます。`0x0`を保留しています。
    - 「ユーザーID」はアプリ側でユーザーを識別するIDです
...
paths:
...
  /api/coconiko/voting/MintProposal:
    post:
      tags:
        - Governance Token
      summary: 新しい投票提案を作成（未実装）
      description: 新しい投票提案を作成
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - proposalId
                - title
                - options
                - endDate
              properties:
                proposalId:
                  type: string
                  description: 提案の一意な識別子
                title:
                  type: string
                  description: 投票タイトル
                description:
                  type: string
                  description: 投票の詳細説明
                options:
                  type: array
                  items: 
                    type: string
                  minItems: 2
                  maxItems: 10
                  description: 投票選択肢
                endDate:
                  type: string
                  format: date-time
                  description: 投票締切日時(UTC)
      responses:
        '201':
          description: 投票提案作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VotingProposal'
  /api/coconiko/voting/MintVoteNFT:
    post:
      tags:
        - Governance Token
      summary: 投票権NFTの発行（未実装）
      description: 有権者に投票権NFTを発行
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - proposalId
              properties:
                proposalId:
                  type: string
                  description: 対象の提案ID
                voterId:
                  type: string
                  description: 有権者アカウントID（デフォルトはリクエストユーザー）
      responses:
        '200':
          description: NFT発行成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VotingNFT'
  /api/coconiko/voting/Cast:
    post:
      tags:
        - Governance Token  
      summary: NFTを使用して投票を実行（未実装）
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - proposalId
                - selectedOption
                - nftTokenId
              properties:
                proposalId:
                  type: string
                selectedOption:
                  type: string
                  description: 選択したオプションのインデックス（0ベース）
                nftTokenId:
                  type: string
                  description: 使用する投票権NFTのトークンID
      responses:
        '200':
          description: 投票成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  txHash:
                    type: string
                    description: ブロックチェーン上のトランザクションハッシュ
  /api/coconiko/voting/Results/{proposalId}:
    get:
      tags:
        - Governance Token
      summary: 投票結果の取得（未実装）
      parameters:
        - in: path
          name: proposalId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 投票結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VotingResult'
  /api/coconiko/queryAssetsWithPagination:
    post:
      tags:
        - Debug
      summary: アセットの取得（ページングあり）
      description: |
        クエリ文字列に基づいてページングされたアセットを返す。詳細は、[CouchDB 検索方法](https://docs.couchdb.org/en/stable/api/database/find.html#find-selectors)を参照。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
                - pageSize
              properties:
                query:
                  type: object
                  description: CouchDBクエリ文字列（JSON形式）
                  example: {
                    "selector": {
                      "docType": "coconiko-coin",
                      "burned": false
                    }
                  }
                pageSize:
                  type: integer
                  description: 1ページあたりのアイテム数
                  minimum: 1
                  maximum: 100000
                  example: 100
                bookmark:
                  type: string
                  description: ページング用のブックマーク（前回レスポンスのmetadata.Bookmark）
                  example: ""
      security:
        - ApiKeyAuth: []
          userId: []
      responses:
        '200':
          description: 正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      metadata:
                        type: object
                        properties:
                          Bookmark:
                            type: string
                            description: 次のページ取得用ブックマーク
                          RecordsCount:
                            type: integer
                            description: 取得レコード数
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            Key:
                              type: string
                              description: アセットキー
                            Record:
                              type: object
                              description: アセットレコード
                example:
                  result: {
                    "metadata": {
                      "Bookmark": "g1AAAABheJzLYWBgYMpgSmHgKy5JLCrJTq2MT8lPzkzJBYqzJRYXpSgY2CqY2P____9_VkKSnH4Gk5AxA4gZgCRQAwCqVRFO",
                      "RecordsCount": 2
                    },
                    "results": [
                      {
                        "Key": "COIN_001",
                        "Record": {
                          "docType": "coconiko-coin",
                          "amount": 100,
                          "owner": "x509::/OU=client/CN=user1.coconiko...",
                          "burned": false
                        }
                      },
                      {
                        "Key": "COIN_002",
                        "Record": {
                          "docType": "coconiko-coin",
                          "amount": 200,
                          "owner": "x509::/OU=client/CN=user2.coconiko...",
                          "burned": true
                        }
                      }
                    ]
                  }
        '400':
          description: 無効なリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/coconiko/postgres/query:
    post:
      tags:
        - Debug
      summary: PostgreSQLクエリの実行（SELECTのみ）
      description: |
        パラメータ化されたPostgreSQL SELECTクエリを実行
        - SELECTクエリのみ許可
        - SQLインジェクション防止のためパラメータ化クエリを使用
        - 生のクエリ結果を返却
      security:
        - ApiKeyAuth: []
          userId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - queryString
              properties:
                queryString:
                  type: string
                  description: SELECTクエリ文（$/<param>/形式のパラメータを使用）
                  example: "SELECT * FROM coconiko.user_info WHERE active = $/active/"
                params:
                  type: object
                  description: クエリパラメータオブジェクト（オプション）
                  example: { "active": true }
      responses:
        '200':
          description: クエリ成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    description: クエリ結果配列
                    items:
                      type: object
                      additionalProperties: true
                example:
                  result: 
                    - { account_id: "0x0", user_id: "system", role: "admin" }
                    - { account_id: "x509...", user_id: "test", role: "user" }
        '400':
          description: リクエストパラメータ検証失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "Bad Request"
                reason: "VALIDATION_ERROR"
                message: "無効なリクエストボディ"
                timestamp: "2024-01-01T00:00:00.000Z"
                errors:
                  - { msg: "有効なSQL SELECT文である必要があります", param: "queryString" }
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
        reason:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: object
...
    VotingProposal:
      type: object
      properties:
        proposalId:
          type: string
        title:
          type: string  
        description:
          type: string
        options:
          type: array
          items: 
            type: string
        totalSupply:
          type: integer
          description: 発行済み投票権数
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    VotingNFT:
      type: object
      properties:
        tokenId: 
          type: string
        proposalId:
          type: string
        voterId:
          type: string
        isUsed:
          type: boolean
          description: 投票済みかどうか
        mintedAt:
          type: string
          format: date-time
    VotingResult:
      type: object
      properties:
        proposalId:
          type: string
        totalVotes:
          type: integer
        results:
          type: object
          additionalProperties:
            type: integer
        invalidVotes:
          type: integer
        calculatedAt:
          type: string
          format: date-time
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: APPのAPIキー
    userId:
      type: apiKey
      in: header
      name: userId
      description: ログインしているユーザーID